🎨 BaseManager
